Move `getCommandClasses()` to `CommandHandlerMethod`
