+++ server/src/test/java/org/spine3/server/aggregate/AggregateRepositoryShould.java
@@ -58,7 +58,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
-import static org.spine3.server.command.CommandHandlingEntity.getCommandClasses;
+import static org.spine3.server.reflect.CommandHandlerMethod.getCommandClasses;
 import static org.spine3.testdata.TestCommandContextFactory.createCommandContext;
 import static org.spine3.validate.Validate.isDefault;
 import static org.spine3.validate.Validate.isNotDefault;
